general: pass transients through arguments
access through global variables has poor performance (?) as the compiler
has to reload the variables after a function call.

allocate on the fly, within each search instance, a transient_t struct
representing the game state during search. to allow debug functions
access to transient state information, the lifetime of the struct is
extended beyond the search call to the next invocation of search.