magics: fix combination of left and right shifts
consecutive shifts in opposite directions are not equivalent to an
overall shift of the sum of the number of shifted bits, although the
intention is as such. fix these shifts by redefining them to use the
combined shift amount, so that unnecessary double shifting is avoided.

fix regression introduced in ac9fc85efa2d177e33871bc6d89455ce9ef42794